Modified LogonScript.ps1
I have made some modifications to the Gaz2600 script, so you don't need to run additional scripts for creating user folders and deleting files, logoff scripts, etc. No file is deleted indeed. You need only 1 script sunning as logon script. What is does is to create username.txt files with computer name inside (e.g. PC1) in a common shared hidden folder (It is important to be hidden, because it is with write access). Every 25 secs script touches the files. If the same user is trying to login from a different PC (e.g. PC2) it checks if usernamer.txt exists and checks if PC names match. If they doesn't match and file is newer than 3 mins it displays the warning dialogue, if it is older than 3 mins it replaces the file with PC2 name so it becomes the current user/pc. Simple as that hope you like it!.
If you need the warning dialogue text to be in Cyrillic like me you need to save the file with UTF8-BOM or ANSI encoding.